Refactor more logic into courses module
