+++ odalc/odalc_admin/forms.py
@@ -1,6 +1,6 @@
 from django import forms
 
-from odalc.base.forms import UserRegisterForm
+from odalc.users.forms import UserRegisterForm
 from odalc.users.models import AdminUser
