+++ odalc/teachers/forms.py
@@ -1,7 +1,7 @@
 from django import forms
 
-from odalc.base.forms import UserRegisterForm
 from odalc.courses.models import Course
+from odalc.users.forms import UserRegisterForm
 from odalc.users.models import TeacherUser
 
 from localflavor.us import forms as localflavor_forms
