+++ odalc/urls.py
@@ -16,12 +16,12 @@ from odalc.lib.s3 import SignS3View
 
 urlpatterns = patterns('',
     url(r'^admins/', include('odalc.odalc_admin.urls', namespace='admins')),
-    url(r'^courses/', include('odalc.base.urls', namespace='courses')),
+    url(r'^courses/', include('odalc.courses.urls', namespace='courses')),
     url(r'^students/', include('odalc.students.urls', namespace='students')),
     url(r'^teachers/', include('odalc.teachers.urls', namespace='teachers')),
     url(r'^users/', include('odalc.users.urls', namespace='users')),
-    url(r'^about/', AboutPageView.as_view(), name='about'),
-    url(r'^donate/', DonatePageView.as_view(), name='donate'),
-    url(r'^sign_s3/', SignS3View.as_view(), name='sign_s3'),
+    url(r'^about/$', AboutPageView.as_view(), name='about'),
+    url(r'^donate/$', DonatePageView.as_view(), name='donate'),
+    url(r'^sign_s3/$', SignS3View.as_view(), name='sign_s3'),
     url(r'^$', HomePageView.as_view(), name='home'),
 )
